+++ src/Core/Runtime/ServiceFactory.cs
@@ -49,7 +49,7 @@ namespace HotChocolate.Runtime
         object IServiceProvider.GetService(Type serviceType) =>
             CreateInstance(serviceType);
 
-        private FactoryInfo CreateFactoryInfo(
+        private static FactoryInfo CreateFactoryInfo(
             IServiceProvider services,
             Type type)
         {
@@ -87,7 +87,7 @@ namespace HotChocolate.Runtime
                 services, type, constructors);
         }
 
-        private FactoryInfo GetBestMatchingConstructor(
+        private static FactoryInfo GetBestMatchingConstructor(
             IServiceProvider services,
             Type type,
             ConstructorInfo[] constructors)