+++ src/Core/Types.Tests/SchemaFirstTests.cs
@@ -54,5 +54,130 @@ namespace HotChocolate
                 await executor.ExecuteAsync("{ __schema { description } }");
             result.MatchSnapshot();
         }
+
+        [Fact]
+        public async Task SchemaBuilder_BindType()
+        {
+            // arrange
+            string sourceText = "type Query { hello: String }";
+
+            // act
+            ISchema schema = SchemaBuilder.New()
+                .AddDocumentFromString(sourceText)
+                .BindComplexType<Query>()
+                .Create();
+
+            // assert
+            IQueryExecutor executor = schema.MakeExecutable();
+            IExecutionResult result =
+                await executor.ExecuteAsync("{ hello }");
+            result.MatchSnapshot();
+        }
+
+        [Fact]
+        public async Task SchemaBuilder_BindType_Configure()
+        {
+            // arrange
+            string sourceText = "type Query { hello: String }";
+
+            // act
+            ISchema schema = SchemaBuilder.New()
+                .AddDocumentFromString(sourceText)
+                .BindComplexType<Query1>(c => c
+                    .To("Query")
+                    .Field(t => t.Hello1())
+                    .Name("hello"))
+                .Create();
+
+            // assert
+            IQueryExecutor executor = schema.MakeExecutable();
+            IExecutionResult result =
+                await executor.ExecuteAsync("{ hello }");
+            result.MatchSnapshot();
+        }
+
+        [Fact]
+        public async Task SchemaBuilder_BindType_And_Resolver()
+        {
+            // arrange
+            string sourceText = "type Query { hello: String }";
+
+            // act
+            ISchema schema = SchemaBuilder.New()
+                .AddDocumentFromString(sourceText)
+                .BindComplexType<Query>()
+                .BindResolver<QueryResolver>(c => c
+                    .To<Query>()
+                    .Resolve(f => f.Hello())
+                    .With(r => r.Resolve(default)))
+                .Create();
+
+            // assert
+            IQueryExecutor executor = schema.MakeExecutable();
+            IExecutionResult result =
+                await executor.ExecuteAsync("{ hello }");
+            result.MatchSnapshot();
+        }
+
+        [Fact]
+        public async Task SchemaBuilder_BindType_And_Resolver_NameBind()
+        {
+            // arrange
+            string sourceText = "type Query { hello: String }";
+
+            // act
+            ISchema schema = SchemaBuilder.New()
+                .AddDocumentFromString(sourceText)
+                .BindComplexType<Query>()
+                .BindResolver<QueryResolver>(c => c
+                    .To("Query")
+                    .Resolve("hello")
+                    .With(r => r.Resolve(default)))
+                .Create();
+
+            // assert
+            IQueryExecutor executor = schema.MakeExecutable();
+            IExecutionResult result =
+                await executor.ExecuteAsync("{ hello }");
+            result.MatchSnapshot();
+        }
+
+
+        [Fact]
+        public async Task SchemaBuilder_BindType_And_Resolver_Implicit()
+        {
+            // arrange
+            string sourceText = "type Query { hello: String }";
+
+            // act
+            ISchema schema = SchemaBuilder.New()
+                .AddDocumentFromString(sourceText)
+                .BindResolver<Query>()
+                .Create();
+
+            // assert
+            IQueryExecutor executor = schema.MakeExecutable();
+            IExecutionResult result =
+                await executor.ExecuteAsync("{ hello }");
+            result.MatchSnapshot();
+        }
+
+        public class Query
+        {
+            public string Hello() => "World";
+        }
+
+        public class Query1
+        {
+            public string Hello1() => "World1";
+        }
+
+        public class QueryResolver
+        {
+            public string Resolve(Query query)
+            {
+                return query.Hello() + " with resolver";
+            }
+        }
     }
 }