+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added access to stitched http response headers (e.g. Set-Cookie) [679](https://github.com/ChilliCream/hotchocolate/issues/679)
 - Added helper to add delegation paths to a field.
 - It is now possible to bind .net types explicitly to schema types with SchemaBuilder.New().BindClrType<ClrType, SchemaType>(). [756](https://github.com/ChilliCream/hotchocolate/issues/756)
+- Added support for schema-first bindings on the `SchemaBuilder` API. [781](https://github.com/ChilliCream/hotchocolate/issues/781)
 
 ### Changed